+++ CHANGELOG.md
@@ -6,8 +6,16 @@
 * Integrate confluent-sql module (built for dbt-adpater) for lower level of REST API wrapper for statement deployment
 * Add the `shift_left version --config` flag to get a view of the application configuration of all parameters.
 * Refactor confluent client to cover what is not in confluent-sql
-* Refactor integration tests
+* Refactor integration tests in tests/it
 * Address issue in SQL parser for column content extraction that missed part due to ')' for TIMESTAMP() type.
+* Add a command for project to list tables and DDL files impacted by production DDL modifications.
+	```sh
+	shift_left project list-impacted-tables modified-files.json --project-path $PIPELINES --output-file impacted-files.json
+	```
+* Modify the foundation DDLs for a set of tables. 
+	```sh
+	shift_left project list-impacted-tables modified-files.json --project-path $PIPELINES --output-file impacted-files.json
+	```
 
 ### v0.1.49 2026-03-21